% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Alter}
\alias{Alter}
\title{Alter}
\description{
Alter datasets.
}
\examples{
Alter$new(cars)$
 source()$
 transform(
   type = "aggregate",
   fields = c("speed", "dist"),
   operations = c("mean", "sum"),
   as = c("meanDist", "sumDist")
 )

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Alter$new()}}
\item \href{#method-source}{\code{Alter$source()}}
\item \href{#method-transform}{\code{Alter$transform()}}
\item \href{#method-getRows}{\code{Alter$getRows()}}
\item \href{#method-getView}{\code{Alter$getView()}}
\item \href{#method-get}{\code{Alter$get()}}
\item \href{#method-eval}{\code{Alter$eval()}}
\item \href{#method-finalize}{\code{Alter$finalize()}}
\item \href{#method-print}{\code{Alter$print()}}
\item \href{#method-clone}{\code{Alter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$new(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataset to alter.}

\item{\code{...}}{Options to pass to the \code{DataSet} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source"></a>}}
\if{latex}{\out{\hypertarget{method-source}{}}}
\subsection{Method \code{source()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$source(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Options to pass to the \code{source} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Source the data
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$transform(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Options to pass to the \code{transform} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Transform the data
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRows"></a>}}
\if{latex}{\out{\hypertarget{method-getRows}{}}}
\subsection{Method \code{getRows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$getRows(clean = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clean}}{Whether to remove the view created from
the global context on exit.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get the rows of the altered data
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getView"></a>}}
\if{latex}{\out{\hypertarget{method-getView}{}}}
\subsection{Method \code{getView()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$getView(clean = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clean}}{Whether to remove the view created from
the global context on exit.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get the view object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$get(what, ..., clean = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{What to retrieve, e.g.: \code{rows} or \code{edges}.}

\item{\code{...}}{Arguments to pass to the deserialiser.}

\item{\code{clean}}{Whether to remove the view created from
the global context on exit.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get from the view
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\if{latex}{\out{\hypertarget{method-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$eval(what, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{Method to evaluated, e.g.: \code{getAllRows()}.}

\item{\code{...}}{Arguments to pass to the deserialiser.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Evaluate method on the view
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Clean up
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
