% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{dataset_dependency}
\alias{dataset_dependency}
\title{Dependency}
\usage{
dataset_dependency()
}
\description{
Import dataset dependency to use in browser.
}
\examples{
library(shiny)

ui <- fluidPage(
 dataset_dependency()
)

server <- function(input, output){}

if(interactive())
 shinyApp(ui, server)

}
