% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter.R
\name{functional}
\alias{functional}
\alias{alter}
\title{Alter}
\usage{
alter(data, ..., dataset = NULL, source = NULL, .rows = TRUE, .return = NULL)
}
\arguments{
\item{data}{A data.frame to alter.}

\item{...}{Altering options.}

\item{dataset}{A \code{list} of options to pass to the \code{DataSet}
instantiation.}

\item{source}{A \code{list} of options to pass to the \code{source}
method.}

\item{.rows}{Wheter to return the whole altered object or just the rows.}

\item{.return}{Method to return, \code{"rows"}.}
}
\description{
Alter a dataset using \href{https://github.com/antvis/data-set}{DataSet}.
}
\examples{
alter(
 cars, 
 type = "aggregate", 
 fields = c("speed", "dist"), 
 operations = c("mean", "mean"), 
 as = c("x", "y")
) 

}
